insert, delete, y associate machine del inventory controller

+++ GymTEC/Data_input_models/manage_inventory/Data_input_inventory.cs
@@ -6,7 +6,7 @@
         public string equipment_type { get; set; }    // tipo de equipo
         public string brand { get; set; }              // marca
         public string serial_number { get; set; }      // número de serie
-        public double cost { get; set; }               // costo
+        public int cost { get; set; }               // costo
         public string branch_name { get; set; }        // nombre de sucursal
     }
 }
